build_rootfs.sh: redo lz4 and cpio to extract initramfs.
If we pipe lz4 and cpio to grep and head. Head will quit immediately
if it gets enough lines. And then the pipe breaks. Sometimes it will
cause the cpio failed at the middle. So not all the files are
extracted. We redo it again without any pipe to fix the build failure
on Linaro's CI.

Change-Id: I25f636f45af97bfd56a297b6826b2fcaf041c0b9
Signed-off-by: Ying-Chun Liu (PaulLiu) <paul.liu@linaro.org>