Manually generate ESP headers to work with older scapy versions.
The version of scapy in the VM image understands ESP packets, but
the version in external/scapy does not. aosp/2365728 added code
that uses scapy.ESP, and while the test passes on the VM image,
it does not pass in VTS.

For now, manually construct ESP headers to make the tests pass in
both environments. The proper fix is to update scapy, but that's
a larger effort.

Fix: 263537835
Test: tests pass on pixel kernel
Test: m vts_kernel_net_tests && adb push $ANDROID_PRODUCT_OUT/testcases/vts_kernel_net_tests/arm64/kernel_net_tests_bin /data/local/tmp/kernel_net_tests_bin && adb shell /data/local/tmp/kernel_net_tests_bin xfrm_test
Change-Id: I6d0c71e62181a9c5b7cc0449bce6d8ebde1863d5