build_rootfs.sh: Set MODULES=most for initramfs-tools
The default of MODULES=dep for initramfs-tools doesn't work well,
because the image generation step happens inside QEMU, so the detected
devices may not match the devices we want to run the images on. Just
use "most", which makes the initramfs a bit larger, but will accommodate
nearly any hardware.

Change-Id: I2c71bbac457d1376539ad61e2a1a67f1637245f5