build_rootfs.sh: Improve virtio_pci_*_dev workaround
Over various kernel versions, the virtio_pci_*_dev support was pushed
into a "modern" module, then it was inverted to "legacy". The new
"legacy" code added for 5.19+ nicely checked if the file existed
before trying to load it, which avoids an error message, but since
we do that, we shouldn't have used "|| true" which will mask other
failures.

At the same time, fix the older "modern" path to take the same approach
as "legacy", so now we only load the necessary modules and no errors are
printed or have to be masked.

Change-Id: I851bb25a7130959cc2fdf6a80bdb6d2e5c745b86