build_rootfs.sh: Switch away from cloud kernel
The cloud kernel is just a subset of the full Debian kernel with some
drivers removed, but on a physical server we need some of those drivers,
so rather than have two configs, let's just use the full kernel.

The NVIDIA driver is bumped again because of a compilation issue with
the Linux 6.0 kernel currently in bullseye backports.

Change-Id: I9c6027a50aa18eddf1207d6d3d12684a86f66fc0