build_rootfs.sh: Work around large initrd
The package list files are passed directly to debootstrap, which fetches
them to a cache which is used by the second stage in foreign mode. Since
this cache is pushed into our initrd it can cause us to blow the limit
of 512M. Temporarily move the largest package to a separate manual
stage.

Change-Id: I7761e7b9cf101eb4d2a9ce19581d94d0d1159897