net-test: skip minLTS test on GSI
Testing on GSI basically doesn't make sense as
we're testing new VTS + new GSI + old vendor
(which includes kernel).  As such it is not
surprising if this test fails... and not much
you can do about it.

In practice this test is now more-or-less duplicated
in VTS netd_integration_test's kernel_test.cpp.

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I7a3642a5d3ba79eb6ac51f0bbf0958f4ed2f76b3