Ensure TcpBaseTest class members always exist.
Currently, TcpBaseTest and SockDestroyTcpTest use class members
that are not always initialized. This is awkward because the code
needs to use hasattr() to see if they have been set and has to
call del to unset them.

Initialize them in the constructor instead. This means that they
always exist and do not need to be deleted, simplifying the code.

Test: test-only change
Change-Id: I575c272f5b354107b27e2f78206c2069941b45a1