Support kernels that make SOCK_DESTROY work on FIN_WAIT sockets.
A change has been sent to net to support destroying FIN_WAIT
sockets:

https://lore.kernel.org/netdev/20240814084622.555672-1-kuro@kuroa.me/T/#t

If applied, this change will cause sock_diag_test to fail,
because it expects that SOCK_DESTROY on FIN_WAIT1 sockets does
nothing. Fix the test to handle the newly-proposed behaviour as
well as the current behaviour.

Test: test-only change
Change-Id: If40089365027b4ef98fdcf695aec9404a4c32fd9