net-test: introduce a standalone '464xlat' implementation
This is based on various portions of Android's clatd, ebpf xlat,
BpfMap.h, other header files, and lots of work done over the years
on the side not as a proper part of Android due to difficulty
of integrating it with the Android code base proper (lack of proper
tests).

We'll need a standalone version of this code to enable
better unit testing and development, and hopefully will
eventually be able to integrate this back into
packages/modules/Connectivity's 464xlat logic.

Test: N/A, but 'make' does work
Flag: EXEMPT unused/dead code
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I2351a849170170b2900b3e14b6ca2a3bb6219954